impl validate_block_pre_execution on each chain
