+++ crates/consensus/common/Cargo.toml
@@ -14,8 +14,3 @@ workspace = true
 # reth
 reth-primitives.workspace = true
 reth-consensus.workspace = true
-
-[dev-dependencies]
-reth-storage-api.workspace = true
-rand.workspace = true
-mockall = "0.12"
