+++ crates/ethereum/consensus/Cargo.toml
@@ -17,3 +17,8 @@ reth-primitives.workspace = true
 reth-consensus.workspace = true
 
 tracing.workspace = true
+
+[dev-dependencies]
+reth-storage-api.workspace = true
+rand.workspace = true
+mockall = "0.12"
