+++ crates/node-core/Cargo.toml
@@ -33,7 +33,6 @@ reth-discv5.workspace = true
 reth-net-nat.workspace = true
 reth-engine-primitives.workspace = true
 reth-tasks.workspace = true
-reth-consensus-common.workspace = true
 reth-beacon-consensus.workspace = true
 reth-prune-types.workspace = true
 reth-stages-types.workspace = true
