+++ crates/optimism/node/src/lib.rs
@@ -28,4 +28,6 @@ pub use reth_optimism_payload_builder::{
     OptimismBuiltPayload, OptimismPayloadBuilder, OptimismPayloadBuilderAttributes,
 };
 
+pub use reth_optimism_consensus::OptimismBeaconConsensus;
+
 pub use reth_evm_optimism::*;
